[FIX] Fixed an issue with unsigned and unsigend long for very large texts.
git-svn-id: http://svn.seqan.de/seqan/trunk@14392 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ core/include/seqan/index/index_fm_prefix_sum_table.h
@@ -89,13 +89,13 @@ struct GetValue<PrefixSumTable<TChar, TSpec> const>
 template <typename TChar, typename TSpec>
 struct Fibre<PrefixSumTable<TChar, TSpec>, FibreEntries>
 {
-    typedef String<unsigned> Type;
+    typedef String<unsigned long> Type;
 };
 
 template <typename TChar, typename TSpec>
 struct Fibre<PrefixSumTable<TChar, TSpec> const, FibreEntries>
 {
-    typedef String<unsigned> const Type;
+    typedef String<unsigned long> const Type;
 };
 
 // ----------------------------------------------------------------------------
